Add files via upload
This code will log whether a device is connected to your local Wifi network.
To run it must be given root permission, so on a pi run $ sudo python3 /address/of/file/location/Wifi_MAC_logger_basic.py 123.csv
The 123.csv  file is just specifying where to log all of the data. If you don't have a csv file already this will create one for you.
The code has been heavily commented to aid beginners and to talk them through every little step. 
You will have to install some modules. Please not that it is difficult to run on a windows device as it is hard to install nmap. I would highly recommend running it on a raspberry pi (I run it on a pi zero W).